fan/flip: move common to algorithm
